Tackling with scaling problems
The LockOptions.UPGRADE with sorted by ID dvds sets ensures that deadlocks will not happen, but at a heavy cost. To be handle the load_test, the java EE container (wildfly here) must be provided with a larger JDBC connection poolsize (5000 seams enough) than the default value. It can be changed with the admin console.